requirements.txt
streamlit==1.27.0
pandas==2.0.3
openai==0.28.0
python-dotenv==1.0.0
openpyxl==3.1.2
